move networks file to db
