+++ src/crawler/crawl.ts
@@ -2,9 +2,9 @@ import moment, { Moment } from 'moment'
 import { encodeNodePublic } from 'ripple-address-codec'
 
 import { query, saveNode } from '../shared/database'
+import { Network } from '../shared/database/networks'
 import { Crawl } from '../shared/types'
 import logger from '../shared/utils/logger'
-import { Network } from '../shared/utils/networks'
 
 import crawlNode from './network'
